Tweaks and improvements

+++ cookbooks/node/recipes/default.rb
@@ -3,6 +3,7 @@ nodejs_tar = "node-v#{current_version}-linux-x64.tar.gz"
 nodejs_tar_dir = nodejs_tar.gsub(/.tar.gz/,'')
 nodejs_bin_url = "http://nodejs.org/dist/v#{current_version}/#{nodejs_tar}"
 destination_dir = "/usr/local"
+install_dirs = ['lib','bin','share']
 
 # want to check against lock and skip entire process if current version matches installed
 
@@ -16,24 +17,16 @@ execute "extract_node_binaries" do
   command "tar xf /usr/local/src/#{nodejs_tar}; mv /#{nodejs_tar_dir} /usr/local/src/"
   user "root"
   action :run
+  not_if "test -f #{Chef::Config[:file_cache_path]}/node_lock"
 end
 
-execute "sparse_copy_node_bin_into_usr_local" do
-  command "rsync -a /usr/local/src/#{nodejs_tar_dir}/bin/ /usr/local/bin/;"
-  user "root"
-  action :run
-end
-
-execute "sparse_copy_node_lib_into_usr_local" do
-  command "rsync -a /usr/local/src/#{nodejs_tar_dir}/lib/ /usr/local/lib/;"
-  user "root"
-  action :run
-end
-
-execute "sparse_copy_node_share_into_usr_local" do
-  command "rsync -a /usr/local/src/#{nodejs_tar_dir}/share/ /usr/local/share/;"
-  user "root"
-  action :run
+install_dirs.each do |dir|
+  execute "sparse_copy_node_#{dir}_into_local" do
+    command "rsync -a /usr/local/src/#{nodejs_tar_dir}/#{dir}/ /usr/local/#{dir}/"
+    user "root"
+    action :run
+    not_if "test -f #{Chef::Config[:file_cache_path]}/node_lock"
+  end
 end
 
 file "#{Chef::Config[:file_cache_path]}/node_lock" do
